pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chronos"
version = "2.0.0"
description = "Autonomous Data Incident Root Cause Analysis Agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "litellm>=1.40.0",
    "httpx>=0.27.0",
    "graphiti-core>=0.3.0",
    "langfuse>=2.38.0",
    "traceloop-sdk>=0.18.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "prov>=2.0.0",
    "python-dotenv>=1.0.0",
    "aioredis>=2.0.0",
    "sse-starlette>=2.1.0",
    "slack-sdk>=3.29.0",
    "mcp>=1.0.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
eval = [
    "deepeval>=0.21.0",
    "ragas>=0.1.0",
]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["chronos"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
